When parsing links try detecting new open braces
Links like:
![](http://www.broadgate.co.uk/Content/Upload/DetailImages/Cyclus700(1).jpg)"
are not correctly parsed because the closing brace of (1) is seen as the
end of the link, which it isn't.

Add code that detects opening braces and tries to find a matching
closing one. If we are left with an uneven pair, fail the link
detections.

Add tests from blackfriday issue #116
https://github.com/russross/blackfriday/issues/116