Add a way to group figures
Figure can be wrapped in figure block quote, where they will be grouped
together as one logical <figure> (in the XML output).

Typical use:

F>  {type="ascii-art"}
F>      +-----+
F>      | ART |
F>      +-----+
F>  Figure: This caption is ignored in v3, but used in v2.
F>
F>  ``` c
F>  printf("%s\n", "hello");
F>  ```
F>
Figure: Caption you will see, for both figures in v3 (in v2 it's ignored).

In v2 this is not supported so the above will result in two figures. Note that in v2
the inner captions *are* used, but the other one will be discarded as there is no
way to typeset it.