Use just 'Abbrev' and allows for plural
We can just use Abbrev under Organization, because toml knows what
it is parsing at that point. So this is no different then parsing
the Abbrev in the main part.

Also use streets, regions, countries etc to speficy multiline
streets, regions, countries etc. This is backward compatible and
most people won't need it.