Fix Windows-style newlines, harder
This mainly adds a re-run of all reference tests with Unix newlines
replaced with Windows ones. Plus, fixes for a bazillion of problems that
this test uncovered.

Node now contains a ContainsWindowsNewlines field, which is set to true
when the contents of the node (either in .Literal or in .content)
include text having '\r' characters. This is then checked at render time
for replacement back to Unix newlines.